Add MIDI functionality with device controller and settings pages
- Implement DeviceControllerPage to manage MIDI device interactions, including sending and receiving MIDI messages.
- Create MidiSettingsPage for scanning and connecting to MIDI devices, handling Bluetooth permissions, and displaying device information.
- Introduce PlayPage with an interactive piano interface, allowing users to play notes and visualize MIDI activity.
- Enhance MIDI data handling and user feedback throughout the application.
